+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Added new spells to move array items up, down, or to a specific row.
 * Implemented copy and paste in the UV editor. UV data is stored on the clipboard in a text format similar to OBJ files, with an additional vertex number field when copying a selection. When pasting a selection to a selection, vertex numbers are remapped, otherwise they are not changed. The paste operation currently does not support the undo/redo feature of the UV editor.
-* Sanitize before save has been added as an option when processing multiple NIF files.
+* Sanitize before save and Remove Bogus Nodes have been added as options when processing multiple NIF files.
 * "Block/Convert" now sets the number of partitions on conversion from NiSkinInstance to BSDismemberSkinInstance. Note that body parts and flags are still initialized with the defaults.
 * PNG format screenshots no longer use the silhouette method for alpha, the output image should be identical to DDS.
 * Fixed BSTriShape vertex positions and bitangent X being lost on changes to the full precision flag.